requirements.txt
fastapi
uvicorn
jinja2
python-multipart
python-dotenv
starlette
requests
openai
